integrated Gender and Object Recognition Services with the API
