--- Recognizer/Recognizer.py
@@ -1,19 +1,0 @@
-#
-# @Recognizer.py Copyright (c) 2021 Jalasoft. # 2643 Av Melchor Perez de Olguin, Colquiri Sud, Cochabamba, Bolivia. # <add direccion de jala la paz>
-# All rights reserved. #
-# This software is the confidential and proprietary information of
-# Jalasoft, ("Confidential Information"). You shall not
-# disclose such Confidential Information and shall use it only in
-# accordance with the terms of the license agreement you entered into
-# with Jalasoft.
-#
-
-from abc import ABC, abstractmethod
-
-
-class Recognizer(ABC):
-
-    @abstractmethod
-    def recognize(self, image_path: str, confidence_threshold: float = 0.1, word: str = None):
-        pass
-
